# docs/trip_planner_api_specification.yaml
...
openapi: 3.0.3
info:
  title: Agentic Trip Planning API
  description: |
    Comprehensive RESTful API for an AI-powered trip planning application.
    Supports user management, trip creation, AI-powered itinerary generation,
    flight/accommodation booking integration, and real-time planning updates.
  version: 1.0.0
  contact:
    name: API Support
    email: api@tripplanner.com
servers:
  - url: https://api.tripplanner.com/v1
    description: Production server
  - url: https://staging-api.tripplanner.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development

# =============================================================================
# SECURITY SCHEMES
# =============================================================================
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoints

  # ===========================================================================
  # REUSABLE SCHEMAS
  # ===========================================================================
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "usr_12345678-1234-1234-1234-123456789abc"
        email:
          type: string
          format: email
          example: "user@example.com"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        avatar:
          type: string
          format: uri
          example: "https://cdn.tripplanner.com/avatars/usr_123.jpg"
        phone:
          type: string
          example: "+1-555-123-4567"
        dateOfBirth:
          type: string
          format: date
          example: "1990-05-15"
        nationality:
          type: string
          example: "US"
        preferredCurrency:
          type: string
          example: "USD"
        timezone:
          type: string
          example: "America/New_York"
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
...
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      description: Create a new user account with email and password
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                firstName:
                  type: string
                lastName:
                  type: string
              required:
                - email
                - password
                - firstName
                - lastName
            example:
              email: "john.doe@example.com"
              password: "SecurePass123!"
              firstName: "John"
              lastName: "Doe"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
                      expiresIn:
                        type: integer
              example:
                user:
                  id: "usr_12345678-1234-1234-1234-123456789abc"
                  email: "john.doe@example.com"
                  firstName: "John"
                  lastName: "Doe"
                  createdAt: "2024-01-15T10:30:00Z"
                tokens:
                  accessToken: "eyJhbGciOiJIUzI1NiIs..."
                  refreshToken: "eyJhbGciOiJIUzI1NiIs..."
                  expiresIn: 3600
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: User login
      description: Authenticate user and return access tokens
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
            example:
              email: "john.doe@example.com"
              password: "SecurePass123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
                      expiresIn:
                        type: integer
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /trips:
    get:
      summary: List user trips
      description: Get all trips for the authenticated user
      tags:
        - Trips
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, planning, planned, active, completed, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of trips retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trip'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create a new trip
      description: Create a new trip with basic information
      tags:
        - Trips
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                destination:
                  $ref: '#/components/schemas/Destination'
                dates:
                  $ref: '#/components/schemas/TripDates'
                travelers:
                  type: integer
                  default: 1
                budget:
                  $ref: '#/components/schemas/Budget'
                preferences:
                  $ref: '#/components/schemas/TripPreferences'
              required:
                - name
                - destination
                - dates
            example:
              name: "Japan Adventure 2024"
              description: "Two-week exploration of Japan's cultural highlights"
              destination:
                type: "country"
                name: "Japan"
                code: "JP"
              dates:
                startDate: "2024-04-01"
                endDate: "2024-04-15"
                isFlexible: false
              travelers: 2
              budget:
                total: 5000
                currency: "USD"
              preferences:
                dietaryRestrictions: ["vegetarian"]
                interestCategories: ["culture", "food", "history"]
      responses:
        '201':
          description: Trip created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
  /trips/{tripId}:
    parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get trip details
      description: Retrieve detailed information about a specific trip
      tags:
        - Trips
      responses:
        '200':
          description: Trip details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '404':
          description: Trip not found

    put:
      summary: Update trip
      description: Update trip information
      tags:
        - Trips
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                dates:
                  $ref: '#/components/schemas/TripDates'
                travelers:
                  type: integer
                budget:
                  $ref: '#/components/schemas/Budget'
                preferences:
                  $ref: '#/components/schemas/TripPreferences'
      responses:
        '200':
          description: Trip updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'

    delete:
      summary: Delete trip
      description: Delete a trip and all associated data
      tags:
        - Trips
      responses:
        '204':
          description: Trip deleted successfully
...
  /trips/{tripId}/flights:
    parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get trip flights
      description: Get all flights for the trip
      tags:
        - Flights
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [outbound, return, internal, all]
            default: all
        - name: status
          in: query
          schema:
            type: string
            enum: [suggested, selected, booked, cancelled]
      responses:
        '200':
          description: Flights retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Flight'

    post:
      summary: Add flight to trip
      description: Manually add a flight to the trip
      tags:
        - Flights
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flight'
      responses:
        '201':
          description: Flight added successfully
...
  /trips/{tripId}/flights/{flightId}/book:
    post:
      summary: Book flight
      description: Mark flight as booked with booking details
      tags:
        - Flights
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
        - name: flightId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookingDetails:
                  $ref: '#/components/schemas/BookingDetails'
      responses:
        '200':
          description: Flight marked as booked

  # ===========================================================================
  # ACCOMMODATION MANAGEMENT
  # ===========================================================================
  /trips/{tripId}/accommodations:
    parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get trip accommodations
      description: Get all accommodations for the trip
      tags:
        - Accommodations
      parameters:
        - name: cityId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [suggested, selected, reserved, booked, cancelled]
      responses:
        '200':
          description: Accommodations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Accommodation'

    post:
      summary: Add accommodation
      description: Manually add an accommodation to the trip
      tags:
        - Accommodations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Accommodation'
      responses:
        '201':
          description: Accommodation added successfully
...
  /trips/{tripId}/accommodations/{accommodationId}/book:
    post:
      summary: Book accommodation
      description: Mark accommodation as booked
      tags:
        - Accommodations
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
        - name: accommodationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookingDetails:
                  $ref: '#/components/schemas/BookingDetails'
      responses:
        '200':
          description: Accommodation marked as booked

  # ===========================================================================
  # ITINERARY MANAGEMENT
  # ===========================================================================
  /trips/{tripId}/itinerary:
    parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get full itinerary
      description: Get the complete trip itinerary
      tags:
        - Itinerary
      responses:
        '200':
          description: Itinerary retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Itinerary'

    put:
      summary: Update full itinerary
      description: Replace the entire trip itinerary
      tags:
        - Itinerary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Itinerary'
      responses:
        '200':
          description: Itinerary updated successfully
...
  /trips/{tripId}/itinerary/items/{itemId}/delay:
    post:
      summary: Delay itinerary item
      description: Move an itinerary item to a different day
      tags:
        - Itinerary
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetDate:
                  type: string
                  format: date
                  description: "New date for the item"
                targetTime:
                  type: string
                  format: time
                  description: "New time for the item (optional)"
                insertAfter:
                  type: string
                  description: "Item ID to insert after (optional)"
              required:
                - targetDate
            example:
              targetDate: "2024-04-03"
              targetTime: "14:00"
      responses:
        '200':
          description: Item delayed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  originalItem:
                    $ref: '#/components/schemas/ItineraryItem'
                  newItem:
                    $ref: '#/components/schemas/ItineraryItem'
                  affectedItems:
                    type: array
                    items:
                      $ref: '#/components/schemas/ItineraryItem'
...